+++ src/main/java/frc/robot/RobotContainer.java
@@ -9,36 +9,25 @@ package frc.robot;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.RunCommand;
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.Constants.OIConstants;
-import frc.robot.autoRoutines.ShootCommandGroup;
-import frc.robot.autoRoutines.ShootWithTurnCommandGroup;
-import frc.robot.commands.AllInCommand;
 import frc.robot.commands.AllOutCommand;
-import frc.robot.commands.climbing.ClimberDownCommand;
-import frc.robot.commands.climbing.ClimberUpCommand;
-import frc.robot.commands.drivetrain.DriveForDistanceCommand;
-import frc.robot.commands.drivetrain.ResetDriveEncodersCommand;
-import frc.robot.commands.drivetrain.StopDriveCommand;
-import frc.robot.commands.drivetrain.TurnNoPIDCommand;
 import frc.robot.commands.drivetrain.TurnToTargetPIDCommand;
 import frc.robot.commands.groups.AimToShootCommandGroup;
 import frc.robot.commands.hood.MoveHoodCommand;
-import frc.robot.commands.hood.ResetHoodEncoderCommand;
-import frc.robot.commands.hopper.HopperOutCommand;
-import frc.robot.commands.hopper.StopFunnelCommand;
-import frc.robot.commands.intake.*;
-import frc.robot.commands.neck.MoveDownNeckCommand;
-import frc.robot.commands.neck.StopNeckCommand;
+import frc.robot.commands.intake.ToggleIntakeCommand;
 import frc.robot.commands.neck.TowerClearCommand;
 import frc.robot.commands.shifting.ToggleShiftingCommand;
-import frc.robot.commands.neck.MoveUpNeckCommand;
-import frc.robot.commands.hopper.HopperInCommand;
-import frc.robot.commands.shooting.*;
-import frc.robot.subsystems.*;
+import frc.robot.commands.shooter.ToggleShooterCommand;
+import frc.robot.subsystems.ClimberSubsystem;
+import frc.robot.subsystems.DrivetrainSubsystem;
+import frc.robot.subsystems.HoodSubsystem;
+import frc.robot.subsystems.HopperSubsystem;
+import frc.robot.subsystems.IntakeSubsystem;
+import frc.robot.subsystems.NeckSubsystem;
+import frc.robot.subsystems.ShiftingGearsSubsystem;
+import frc.robot.subsystems.ShooterSubsystem;
 import frc.robot.utils.Limelight;
 //import frc.robot.utils.DoubleButton;
