+++ src/main/java/frc/robot/autoRoutines/ShootCommandGroup.java
@@ -6,7 +6,7 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commands.hopper.HopperInCommand;
 import frc.robot.commands.intake.IntakeInCommand;
 import frc.robot.commands.neck.MoveUpNeckCommand;
-import frc.robot.commands.shooting.StartSpinCommand;
+import frc.robot.commands.shooter.StartSpinCommand;
 import frc.robot.subsystems.*;
 
 public class ShootCommandGroup extends SequentialCommandGroup {
