+++ src/main/java/frc/robot/autoRoutines/TurnThenForwardThenShootCommandGroup.java
@@ -2,7 +2,7 @@ package frc.robot.autoRoutines;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commands.drivetrain.DriveForDistanceCommand;
-import frc.robot.commands.shooting.StopSpinCommand;
+import frc.robot.commands.shooter.StopSpinCommand;
 import frc.robot.subsystems.DrivetrainSubsystem;
 import frc.robot.subsystems.HopperSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
