+++ src/main/java/frc/robot/commands/shooter/ReverseShootCommand.java
@@ -1,15 +1,13 @@
-package frc.robot.commands.shooting;
+package frc.robot.commands.shooter;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants.ShooterConstants;
 import frc.robot.subsystems.ShooterSubsystem;
 
 
-public class StartSpinCommand extends CommandBase {
-
+public class ReverseShootCommand extends CommandBase {
     private final ShooterSubsystem shooterSubsystem;
 
-    public StartSpinCommand(ShooterSubsystem shooterSubsystem) {
+    public ReverseShootCommand(ShooterSubsystem shooterSubsystem) {
         this.shooterSubsystem = shooterSubsystem;
         addRequirements(shooterSubsystem);
     }
@@ -21,16 +19,17 @@ public class StartSpinCommand extends CommandBase {
 
     @Override
     public void execute() {
-        shooterSubsystem.shoot();
+        shooterSubsystem.reverseShoot();
     }
 
     @Override
     public boolean isFinished() {
-        return Math.abs(shooterSubsystem.getSpeed()) >= ShooterConstants.SHOOT_1_SPEED;
+        // TODO: Make this return true when this Command no longer needs to run execute()
+        return false;
     }
 
     @Override
     public void end(boolean interrupted) {
-
+        shooterSubsystem.stopShoot();
     }
 }
