+++ src/main/java/frc/robot/commands/shooter/StopSpinCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.shooting;
+package frc.robot.commands.shooter;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ShooterSubsystem;
