+++ src/main/java/frc/robot/commands/shooter/ToggleShooterCommand.java
@@ -1,13 +1,14 @@
-package frc.robot.commands.shooting;
+package frc.robot.commands.shooter;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.subsystems.ShooterSubsystem;
 
 
-public class ReverseShootCommand extends CommandBase {
+public class ToggleShooterCommand extends CommandBase {
     private final ShooterSubsystem shooterSubsystem;
 
-    public ReverseShootCommand(ShooterSubsystem shooterSubsystem) {
+    public ToggleShooterCommand(ShooterSubsystem shooterSubsystem) {
         this.shooterSubsystem = shooterSubsystem;
         addRequirements(shooterSubsystem);
     }
@@ -19,17 +20,23 @@ public class ReverseShootCommand extends CommandBase {
 
     @Override
     public void execute() {
-        shooterSubsystem.reverseShoot();
+        if (ShooterSubsystem.getShooterStatus() == ShooterSubsystem.shooterStatus.OFF) {
+            shooterSubsystem.shoot();
+        } else if (ShooterSubsystem.getShooterStatus() == ShooterSubsystem.shooterStatus.FORWARD){
+            shooterSubsystem.stopShoot();
+        }else {
+            shooterSubsystem.stopShoot();
+        }
+
     }
 
     @Override
     public boolean isFinished() {
         // TODO: Make this return true when this Command no longer needs to run execute()
-        return false;
+        return true;
     }
 
     @Override
     public void end(boolean interrupted) {
-        shooterSubsystem.stopShoot();
     }
 }
