+++ src/main/java/frc/robot/subsystems/ShooterSubsystem.java
@@ -15,8 +15,6 @@ public class ShooterSubsystem extends SubsystemBase {
 
     private TalonSRX shooterLeftTalon;
     private TalonSRX shooterRightTalon;
-//    public Servo hoodServo1;
-//    public Servo hoodServo2;
     public TalonSRX hoodTalon;
     private double distance;
     public enum ShooterStatus{
